add test code for full pipeline
