+++ active_pipeline/call_active.sh
@@ -2,9 +2,10 @@
 
 # Train model and store artifacts in wandb
 
-config_path='/content/lightning-pose/active_loop/configs/config_ibl_active.yaml'
+config_path="${LIGHTPOSE_DIR}/active_loop/configs/config_ibl_active.yaml"
 
-python /content/lightning-pose/active_loop/call_active_loop.py $config_path
+cd ${LIGHTPOSE_DIR}
+python active_loop/call_active_loop.py $config_path
